fix: CSV-Import zerstört keine deutschen Dezimalzahlen mehr (#31)
parseFloat("1,50") liefert 1 — beim CSV-Import wurde aus 1,50 € also
stillschweigend 1 €. Betraf alle Zahlen-Attribute beim deutschen
Komma-Format.

Neuer Helfer parseDecimal (utils/csv.ts) versteht deutsches (1.234,56 /
1,50) und englisches (1,234.56 / 1.50) Format und unterscheidet Dezimal-
von Tausendertrennzeichen heuristisch. Eingesetzt im aktiven Import-Pfad
(CollectionContext.importCSV) und im Helfer parseCategoryCSV.

8 neue Tests decken beide Formate, Tausendertrennzeichen, Ganzzahlen
und ungültige Werte ab. Tests: 97/97 grün.

+++ src/context/CollectionContext.tsx
@@ -721,7 +721,7 @@ export const CollectionProvider: React.FC<CollectionProviderProps> = ({ children
           
           // Wert entsprechend des Datentyps konvertieren
           if (attr.type === 'number') {
-            itemValues[attrId] = value === '' ? 0 : parseFloat(value);
+            itemValues[attrId] = value === '' ? 0 : parseDecimal(value);
             if (isNaN(itemValues[attrId])) {
               results.errors.push(`Zeile ${i+1}, Spalte ${header}: Ungültiger Zahlenwert "${value}"`);
               itemValues[attrId] = 0;
